move TableSchemaGetter to simple schema conflict res class
